+++ Minibank.Core/Services/ICurrencyRate.cs
@@ -4,7 +4,7 @@ namespace Minibank.Core.Services
 {
     public interface ICurrencyRate
     {
-        public double GetCurrencyRate(string currencyCode);
+        public decimal GetCurrencyRate(string currencyCode);
     }
 }
